Move MarketDocument components to new folder and namespace
