+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/MarketDocument/Cim/CodingScheme.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.MarketDocument
+namespace GreenEnergyHub.Charges.Infrastructure.MarketDocument.Cim
 {
     public enum CodingScheme
     {
