+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/MarketDocument/Cim/MarketParticipantRoleMapper.cs
@@ -15,7 +15,7 @@
 using System.ComponentModel;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 
-namespace GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.MarketDocument
+namespace GreenEnergyHub.Charges.Infrastructure.MarketDocument.Cim
 {
     public static class MarketParticipantRoleMapper
     {
