+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Infrastructure/MarketDocument/Cim/IndustryClassificationMapperTests.cs
@@ -14,11 +14,11 @@
 
 using System.ComponentModel;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
-using GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.MarketDocument;
+using GreenEnergyHub.Charges.Infrastructure.MarketDocument.Cim;
 using Xunit;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.Tests.Infrastructure.Messaging.Serialization.MarketDocument
+namespace GreenEnergyHub.Charges.Tests.Infrastructure.MarketDocument.Cim
 {
     [UnitTest]
     public class IndustryClassificationMapperTests
